requirements.txt
torch
torchvision
torchaudio
selenium
requests
Pillow
numpy
tqdm
matplotlib